Fix install, custom ops
