+++ coconut/ipy_endpoints.py
@@ -1,0 +1,76 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+# -----------------------------------------------------------------------------------------------------------------------
+# INFO:
+# -----------------------------------------------------------------------------------------------------------------------
+
+"""
+Endpoints for Coconut's IPython integration.
+"""
+
+# -----------------------------------------------------------------------------------------------------------------------
+# IMPORTS:
+# -----------------------------------------------------------------------------------------------------------------------
+
+from __future__ import print_function, absolute_import, unicode_literals, division
+
+from coconut.root import *  # NOQA
+
+from coconut.constants import coconut_kernel_kwargs
+
+
+# -----------------------------------------------------------------------------------------------------------------------
+# IPYTHON:
+# -----------------------------------------------------------------------------------------------------------------------
+
+def embed(kernel=False, depth=0, **kwargs):
+    """If _kernel_=False (default), embeds a Coconut Jupyter console
+    initialized from the current local namespace. If _kernel_=True,
+    launches a Coconut Jupyter kernel initialized from the local
+    namespace that can then be attached to. _kwargs_ are as in
+    IPython.embed or IPython.embed_kernel based on _kernel_."""
+    from coconut.icoconut.embed import embed, embed_kernel, extract_module_locals
+    if kernel:
+        mod, locs = extract_module_locals(1 + depth)
+        embed_kernel(module=mod, local_ns=locs, **kwargs)
+    else:
+        embed(stack_depth=3 + depth, **kwargs)
+
+
+def load_ipython_extension(ipython):
+    """Loads Coconut as an IPython extension."""
+    # add Coconut built-ins
+    from coconut import __coconut__
+    newvars = {}
+    for var, val in vars(__coconut__).items():
+        if not var.startswith("__"):
+            newvars[var] = val
+    ipython.push(newvars)
+
+    # import here to avoid circular dependencies
+    from coconut import convenience
+    from coconut.exceptions import CoconutException
+    from coconut.terminal import logger
+
+    magic_state = convenience.get_state()
+    convenience.setup(state=magic_state, **coconut_kernel_kwargs)
+
+    # add magic function
+    def magic(line, cell=None):
+        """Provides %coconut and %%coconut magics."""
+        try:
+            if cell is None:
+                code = line
+            else:
+                # first line in block is cmd, rest is code
+                line = line.strip()
+                if line:
+                    convenience.cmd(line, default_target="sys", state=magic_state)
+                code = cell
+            compiled = convenience.parse(code, state=magic_state)
+        except CoconutException:
+            logger.print_exc()
+        else:
+            ipython.run_cell(compiled, shell_futures=False)
+    ipython.register_magic_function(magic, "line_cell", "coconut")
